Feat : StatModifier Upgrade System
유저가 비용을 지불하고 스텟을 업그레이드 하는 시스템 구현

+++ Assets/ScriptableObject/Enemy/EnemySO.cs
@@ -27,5 +27,5 @@ public class EnemySO : ScriptableObject
 {
     [field: SerializeField] public EnemyData EnemyData { get; private set; }
     [field: SerializeField] public EnemyAttackInfoData EnemyAttackInfoData { get; private set; }
-    public int GoldReward;
+    public BigInteger GoldReward;
 }
